Create seconds_per_year.py
Seconds in a year
Use Python to calculate the number of seconds in a year, and tell the user what the result is in a nice print statement! You should use constants for this exercise -- there are 365 days in a year, 24 hours in a day, 60 minutes in an hour, and 60 seconds per minute.